refactor(spec): tighten v4 module boundaries
