+++ crates/coral-spec/src/v4/surfaces/mod.rs
@@ -1,1 +1,6 @@
-pub(crate) mod openapi;
+mod openapi;
+
+pub use openapi::{
+    OpenApiDocumentMetadata, import_openapi_surface, normalize_source_document,
+    openapi_document_metadata,
+};
